@@ -1,21 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Configuration.AzureAppConfiguration;
-
+using Microsoft.Extensions.Hosting;
 
 namespace Web
 {
     public class Program
     {
-        public static IHostingEnvironment HostingEnvironment;
+        public static IWebHostEnvironment HostingEnvironment;
         public static void Main(string[] args)
         {
             try
